refactor: route display name overrides through config pipeline
