+++ src/shared/agent-display-names.test.ts
@@ -315,3 +315,4 @@ describe("#given user-defined display name overrides via config", () => {
     })
   })
 })
+
